Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/child-src/child-src-worker-allowed.sub-expected.txt

This test used to check the child-src csp controlling worker creation. This behaviour has been deprecated but it's still supported until the transition is done. This still tests that behaviour but we need to go through extra hoops to make sure 'script-src' does not affect the result in any way (for instance by allowing 'self').


PASS Worker is allowed because of deprecated 'child-src' directive
